moved InputBuf to buffer module
